Rewrote committeeconfigs to build on viewcollectives
