+++ activity_calendar/committee_pages/config.py
@@ -1,6 +1,6 @@
 from django.urls import path, include, reverse
 
-from committees.config import CommitteeConfig
+from committees.committeecollective import CommitteeConfig
 
 from activity_calendar.committee_pages.views import ActivityCalendarView, AddActivityMomentCalendarView
 
@@ -8,14 +8,13 @@ from activity_calendar.committee_pages.views import ActivityCalendarView, AddAct
 
 class ActivityConfig(CommitteeConfig):
     url_keyword = 'activity'
-    tab_select_keyword = 'tab_activity'
     name = 'Activities'
-    url_name = 'committees:group_activities'
+    url_name = 'group_activities'
     requires_permission = 'activity_calendar.view_activity'
 
     def get_urls(self):
         """ Builds a list of urls """
         return [
-            path('', ActivityCalendarView.as_view(config_class=ActivityConfig), name='group_activities'),
-            path('<int:activity_id>/add/', AddActivityMomentCalendarView.as_view(config_class=ActivityConfig), name='add_activity_moment'),
+            path('', ActivityCalendarView.as_view(config=self), name='group_activities'),
+            path('<int:activity_id>/add/', AddActivityMomentCalendarView.as_view(config=self), name='add_activity_moment'),
         ]
